refactor(workspace): create claudectl-tui crate, move recorder/session_recorder/demo
First step of #275: introduce a separate TUI crate so the terminal UI
can evolve independently of the binary's brain/coord/bus subsystems.
Starts with the standalone peripherals that have no `App` dependency:

* `crates/claudectl-tui/` with its own Cargo.toml, depends only on
  `claudectl-core` + `serde_json`. Registered in the workspace
  `[workspace] members` list.
* Moved `recorder.rs`, `session_recorder.rs`, `demo.rs` into
  `crates/claudectl-tui/src/`. Internal `crate::transcript::*` and
  `crate::session::*` refs rewritten to `claudectl_core::*`.
* Binary's `src/lib.rs` re-exports `pub use claudectl_tui::{demo,
  recorder, session_recorder};` so existing `crate::recorder::*` /
  `crate::demo::*` / `crate::session_recorder::*` paths in app.rs and
  main.rs resolve unchanged.
* `src/main.rs` swaps `mod demo;` / `mod recorder;` / `mod
  session_recorder;` for `use claudectl_tui::{demo, recorder,
  session_recorder};`.

`demo_peers()` couldn't come along — it references
`crate::ui::peers::PeerDisplayInfo`, which still lives in the binary's
`ui/` directory. Extracted to a new tiny binary-side module
`src/demo_peers.rs`; app.rs and main.rs reach it via
`crate::demo_peers::demo_peers`. Once `ui::peers` migrates into
claudectl-tui (follow-up PR), `demo_peers` can join it.

CI: new `TUI (standalone)` job mirrors the existing `Core (standalone)`
job — builds, clippys, and tests `claudectl-tui` with `-p` only, so an
accidental upward dep would fail the workspace-direction guard.

Verified: cargo build / test / clippy / fmt all clean across the full
workspace and standalone for both crates. Layering grep still passes
against the core pattern.

Toward #275 (extract claudectl-tui crate).
