+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
-members = ["crates/claudectl-core"]
+members = ["crates/claudectl-core", "crates/claudectl-tui"]
 # The root package is also a workspace member (the binary crate). Future epics
-# (extract claudectl-tui, claudectl-brain, etc. — see #279) will add more.
+# (e.g. claudectl-brain — see #279) may add more.
 
 [package]
 name = "claudectl"
@@ -35,6 +35,7 @@ path = "src/lib.rs"
 
 [dependencies]
 claudectl-core = { path = "crates/claudectl-core" }
+claudectl-tui = { path = "crates/claudectl-tui" }
 ratatui = "0.29"
 crossterm = "0.28"
 serde = { version = "1", features = ["derive"] }
