+++ crates/claudectl-tui/src/recorder.rs
